agent/pyproject.toml
[project]
name = "alterego-agent"
version = "0.1.0"
description = "AlterEgo Fetch.ai uAgent sidecar — bridges FastAPI to ASI:One Chat Protocol"
requires-python = ">=3.12"
dependencies = [
    "uagents>=0.18",
    "uagents-core>=0.4",
    "httpx>=0.27",
    "python-dotenv>=1.0",
]
